relocate chain & currency types & constants to relevant directories
