+++ packages/storykit/src/components/CollectionGraph/CollectionGraph.tsx
@@ -1,8 +1,8 @@
 import { listResource } from "@/lib/api"
-import { STORYKIT_SUPPORTED_CHAIN } from "@/lib/chains"
 import { cn } from "@/lib/utils"
 import { useStoryKitContext } from "@/providers/StoryKitProvider"
 import { RESOURCE_TYPE } from "@/types/api"
+import { STORYKIT_SUPPORTED_CHAIN } from "@/types/chains"
 import { useQuery } from "@tanstack/react-query"
 import React, { useEffect, useRef, useState } from "react"
 import { LinkObject } from "react-force-graph-2d"
