+++ packages/storykit/src/components/LicenseTermsList/LicenseTermsList.tsx
@@ -1,11 +1,11 @@
 import { noLicenseTerms } from "@/constants/pil-flavors"
 import { cn } from "@/lib"
 import { getResource } from "@/lib/api"
-import { STORYKIT_SUPPORTED_CHAIN } from "@/lib/chains"
 import { convertLicenseTermObject } from "@/lib/functions/convertLicenseTermObject"
 import { useStoryKitContext } from "@/providers/StoryKitProvider"
 import { PILTerms } from "@/types"
 import { RESOURCE_TYPE } from "@/types/api"
+import { STORYKIT_SUPPORTED_CHAIN } from "@/types/chains"
 import { useQuery } from "@tanstack/react-query"
 import { cva } from "class-variance-authority"
 import { CircleCheck, CircleMinus, Info } from "lucide-react"
