+++ packages/storykit/src/components/RoyaltyGraph/__docs__/Example.tsx
@@ -1,4 +1,3 @@
-import { STORYKIT_SUPPORTED_CHAIN } from "@/lib/chains"
 import { getCollectionByAddress, getNFTByWallet } from "@/lib/simplehash"
 import { useStoryKitContext } from "@/providers"
 import { RoyaltyGraphProvider } from "@/providers/RoyaltyGraphProvider/RoyaltyGraphProvider"
@@ -7,6 +6,7 @@ import {
   ILIAD_TESTNET_ROYALTY_GRAPH_IP_ASSETS, // ILIAD_TESTNET_ROYALTY_GRAPH_LAP_ASSETS,
   // ILIAD_TESTNET_ROYALTY_GRAPH_LRP_ASSETS,
 } from "@/stories/data"
+import { STORYKIT_SUPPORTED_CHAIN } from "@/types/chains"
 import React, { FC, useEffect, useState } from "react"
 import { Address } from "viem"
