+++ packages/storykit/src/constants/index.ts
@@ -1,4 +1,2 @@
-export * from "./assets"
-export * from "./simplehash"
 export * from "./chains"
 export * from "./currencies"
