+++ packages/storykit/src/lib/index.ts
@@ -1,4 +1,3 @@
 export * from "./utils"
 export * from "./functions/convertLicenseTermObject"
 export * from "./chains"
-export * from "./currencies"
