+++ packages/storykit/src/providers/RoyaltyGraphProvider/RoyaltyGraphProvider.tsx
@@ -1,5 +1,5 @@
-import { STORYKIT_SUPPORTED_CHAIN } from "@/lib"
 import { getRoyaltiesByIPs } from "@/lib/royalty-graph"
+import { STORYKIT_SUPPORTED_CHAIN } from "@/types/chains"
 import { RoyaltiesGraph } from "@/types/royalty-graph"
 import { useQuery } from "@tanstack/react-query"
 import React from "react"
