+++ packages/storykit/src/providers/StoryKitProvider/__docs__/StoryKitProvider.mdx
@@ -58,14 +58,14 @@ export default ExampleComponent;
 
 The StoryKitProvider provides the following data to its children:
 
-- **chain** - the Story Chain config as defined in [chains.ts](https://github.com/storyprotocol/storykit/blob/main/packages/storykit/src/lib/chains.ts)
+- **chain** - the Story Chain config as defined in [chains.ts](https://github.com/storyprotocol/storykit/blob/main/packages/storykit/src/constants/chains.ts)
 - **viemChain** - the Chain config in viem Chain format
 - **apiKey** - The users API key if provided
 - **appId** - The users appId if provided
 - **theme** - The style theme, 'default' if none provided
 - **mode** - The style mode, 'light', 'dark' or undefined, will default to system if none provided
 - **themeClass** - The theme class to use if user wants to overwrite the theme mode
-- **defaultCurrency** - The default currency as defined in [currencies.ts](https://github.com/storyprotocol/storykit/blob/main/packages/storykit/src/lib/currencies.ts), sUSD if none provided
+- **defaultCurrency** - The default currency as defined in [currencies.ts](https://github.com/storyprotocol/storykit/blob/main/packages/storykit/src/constants/currencies.ts), sUSD if none provided
 
 #### Arguments
